Changes: * Rename getOpninonController to getClientsOpinionsController * Fix bug with get data from databse
